+++ apps/gm-toolkit-session-management-settings.js
@@ -29,7 +29,7 @@ export default class GMToolkitSessionManagementSettings
 
   static PARTS = {
     form: {
-      template: "modules/wfrp4e-gm-toolkit/templates/gm-toolkit-settings.html",
+      template: GMToolkit.modulePath("templates/gm-toolkit-settings.html"),
       classes: ["gmtoolkit", "scrollable"]
     },
     footer: {